+++ ichnaea/models/api.py
@@ -6,7 +6,6 @@ from sqlalchemy import (
 from sqlalchemy.dialects.mysql import (
     INTEGER as Integer,
 )
-from sqlalchemy.orm import load_only
 
 from ichnaea.models.base import _Model
 from ichnaea.models.hashkey import (
@@ -25,7 +24,6 @@ class ApiKey(HashKeyQueryMixin, _Model):
 
     _hashkey_cls = ApiHashKey
     _query_batch = 100
-    _essential_columns = ('valid_key', 'maxreq', 'log', 'shortname')
 
     valid_key = Column(String(40), primary_key=True)
 
@@ -37,23 +35,10 @@ class ApiKey(HashKeyQueryMixin, _Model):
     allow_fallback = Column(Boolean)
     # A readable short name used in metrics
     shortname = Column(String(40))
-    # A contact address
-    email = Column(String(255))
-    # Some free form context / description
-    description = Column(String(255))
 
     @property
     def name(self):
         return self.shortname or self.valid_key
 
     def __str__(self):  # pragma: no cover
-        return self.name
-
-    @classmethod
-    def querykey(cls, session, key):
-        if key is None:  # pragma: no cover
-            return None
-        # by default exclude the long email/description string fields
-        return (session.query(cls)
-                       .filter(*cls.joinkey(key))
-                       .options(load_only(*cls._essential_columns)))
+        return '<ApiKey>: %s' % self.name